Emit notification upon MiqEvent, EmsEvent, or RequestEvent
This is transient solution. Getting the notifications done right will
require more iterations. Here, we introduce minimal, but viable and
extensible increment.

Notifications are hard. It is fairly easy for us to flood the users
with plethora of notifications, it will be much more difficult to build
a system that sends only targeted, useful and contextful messages. Two
problems arise here, (A) for users stuff happens in context, and (B)
users may want to customize conditions upon which the notification is
emitted

(A) Ours EventStream oftentimes lacks context. An event either tells
that something had happened, or that some(body/thing) has requested it
to happen. For a notification to be useful, it is required to combine
both. In a nutshell, we need to put an EmsEvent into context of
preceding MiqEvent (if exists). Chances are that Control and Automate
will benefit from this as well.

(B) We already have MiqAlert subsystem to emit e-mails (or any MiqAction
in general) upon event. MiqAlert offers great customization for users.
This system should be used in next step. However, notifications will
need more customization than MiqAlert offers now. A state machine may be
needed to notify users about complex background tasks.

Fortunately, we don't need to handle all the problems at once. We can
start introduce notifications that we have enough context for, adding
the more difficult later.