Add todo about not unlocking the lockfile in bin/update
This is really slow and should not be needed.

bin/update is to be run very often.
It's not a "reset all the things" script.
If we need such a script, let's call it bin/reset.

We should be running the scripts as such:
bin/setup: once for each new environment
bin/update: often, following git pull
bin/reset: when something is messed up and you don't know what to do.