Use name instead of base_class.name in process_tasks_mixin
The UI calls .invoke_tasks on a class, and we should use that class
for whatever meta-calls we need to make from the mixin.

For example if the UI calls Vm.process_tasks we should not try to look
up the API collection for VmOrTemplate, and we definitiely shouldn't
requeue the message as VmOrTemplate.process_tasks if we fail.