Convert `#` based headers to underline style
Switching back for a few reasons

* Works better with tools like `hub` that use `#` as the comment
  character (technically configurable so this is avoidable)
* Arguably looks better (but that is on a case by case basis)

More so for the first reason above.  The reason for them being switched
originally was because we currently use the `##` syntax in other
markdown documents more so than the `---`, as many developers prefer
less keystrokes when typing out the headers.

If you are considering on changing this back, please consider opening up
this as a discussion/issue first to gain a consensus from the group to
avoid too much churn with this.