Don't auto-reload ConsoleMethods
Fixes a bug in our Rails console commands that's been making people a
kind of crazy town banana pants.

```
> simulate_queue_worker  (works)
> reload! (works)
> simulate_queue_worker (fails)

A copy of Vmdb::ConsoleMethods has been removed from the module tree but is still active!
```

The ConsoleMethods module that defines these commands is in
autoload_paths, but you cannot auto-reload ConsoleMethods as these are
mixed in to the TOPLEVEL_BINDING for pry/irb. So only load that once.