Pass in the correct delay_for_action depending on the action
The service#start action requires a delay for first action (including all subsequent queued actions.)
The #stop action requires the same, except we pass in the delay_action for the last index as opposed to the first
for the start action.

This allows the first service in each bundle to pass the correct delay as opposed to defaulting to 0 in all cases.

https://www.pivotaltracker.com/story/show/128461221