Remove #fetch_value from API helper
This may have been needed to pass to a method a reference to a `let`
(as a symbol) while delaying its evaluation until needed. AFAICT this is
no longer the case and it is more straightforward to always pass the
value, avoiding the need to conditionally do `public_send`s with symbols
that get passed around.