First cut of tenant inheritance for sharing
Enable VmOrTemplate because we can't test both direct and indirect tenant
shares with MiqTemplate.  MiqTemplate has a TENANT_ACCESS_STRATEGY in
Rbac::Filterer of :ancestor_ids (so children tenants can see ancestor templates).

Since Vm has :descendant_ids as the strategy, we won't see ancestor Vms in Rbac,
so it is better to test both direct and indirect tenant sharing with it.

https://www.pivotaltracker.com/story/show/131005337