Avoid constantize in #name_for_klass
`CollectionConfig#name_for_klass` will call `#constantize` on every value
for `:klass` in the API config. This can cause an issue with autoloading
if a class included in the config calls this method as part of its
definition - a circular dependency resulting in a `LoadError`. This can
be avoided if we instead inspect the name of the class passed in and
compare that to the config, removing the need to call `#constantize` at
all.