Refactor prefix out of tag_list
It is always the same.

This move is not intuitive. Even though it comes with another iteration
over tags it allows us to split reconstruction of tag_list on the entity
with a prefix that determines the rates.

Please consult ChargebackContainerImage.where_clause to verify that the
following are the equivalents:
 * Chargeback.report_cb_model(self.class.name).underscore
 * container_image if perf.resource_type == Container.name