Move the module extend to the subclass
This ensures that we will properly prepend our intercept classes
to route method calls.

In the case of methods defined in the 'Guest' Vm concern, the
classes we define, which were prepended to VmOrTemplate, would actually
come after the included concern which defines the methods we want to hide in
the ancestors chain.

If we prepend to Vm we are sure to come before any modules included
by Vm.