Unless we're retrying, queue non-existing authentication checks
We need to skip requeues when there is already the same attempt on the
queue.

Unfortunately, this is always the case when we're "delivering" an
authentication check message and it fails for a retryable error.

Therefore, we need to force the "queueing" of the retry.  We accomplish
this by passing a force option from the "retryable error" scenario.