Don't requeue attempt 6 or more.
authentication_check_retry_deliver_on was returning nil for attempt 6 or
more.  This would cause us to queue a message with a nil deliver_on.

We move the (attempt < MAX_ATTEMPTS) knowledge out of this method to the
caller, authentication_check_types_queue, which is in better position
to skip the queueing.