Replace api_auth_token factory with unused auth_token
The auth_token factory is unused in the test code and serves only as a
kind of 'abstract' factory for that of api_auth_token. Furthermore, the
attributes supplied in the latter are illustrative but not necessary to
build a valid AuthToken object. Any attributes necessary for existing
tests to pass are better expressed in the tests themselves.