Enables travis caches for the bower_components dir
For this, bundler is still cached through the default mechanisms, but in
addition to that, caching the vendor/assets/bower_components directory
as well.

To make sure the cache is not out of date, the bower.json file is
cmp'd (compared, using the `cmp` command) to the one that was cached
in vendor/assets/bower_components directory.  This is added to the cache
in the `before_cache` step.  Unless this passes, the directory is
removed and bower is run and updated normally.

In this, bower is also skipped from running all together if the assets
already exist.  So the only way that the directory should exist is if
the bower.json file matches the one in the directory (it also has to
exist for this to work), and there has been a successful run in the past
that uploaded the cached assets, otherwise bower is run normally.