Support hash values in dropdown orchestration dialog fields
Allowed values passed to the
`OrchestrationTemplate::OrchestrationParameter` are a way of limiting
the user to a predefined list of supported values. These values are
presented in a dropdown whose values are the same as the text
descriptions.

In order to allow specifying these values as specific keys (IDs), this
patch adds support for hash object representing the allowed values. In
this case, the dropdown will use hash keys for values and hash values
for texts displayed to the user.

This patch does not break legacy use and adds a set of specific tests
for different cases.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>