Custom supports methods shall not crash UI
SupportsFeatureMixin is used in many places in the application. For
dynamic features a CustomFeatureMixin allows blocks like the following.

    supports :x do
      provider.api_version.abilities.include? :x
    end

The problem occurs when an exception happens in the given block. As it
does in https://bugzilla.redhat.com/show_bug.cgi?id=1394021

The supports methods are used in the UI and they are not expected to
raise exceptions. The above mentioned bug crashes the vm details page.

Given our direction wrt code decentralization and pluggable features, I
am afraid, our chances to keep custom supports blocks sane are
deteriorating. That's why I advise for rescue.