Add a migration to remove settings_changes rows for the excludes
Because we have removed the excludes from the settings.yml file
we should remove the changes made in the database.

This will leave current installations of pglogical in the same state
they were in (we will read the excludes from the pglogical tables),
but will reset new installations created when moving from rubyrep
to pglogical.

https://bugzilla.redhat.com/show_bug.cgi?id=1389821