Add support for hiding resources
The requester may only want the count/subcount from a query. If the
number of resources to be rendered is very large, iterating over them
on the backend unnecessarily carries a performance penalty, as well as
generating a large response body to go over the network. Adding an
option to hide the resources addresses both issues.